requirements.txt
python-dotenv
tiktoken
google-generativeai
prettytable
openai
pandas
ipywidgets
tqdm
matplotlib
seaborn
scipy
